# main.go
package main

import (
	"encoding/json"
	"net/http"

	_ "github.com/gmelodie/buggy1/docs"

	logging "github.com/ipfs/go-log/v2"
	"github.com/labstack/echo/v4"
	echoSwagger "github.com/swaggo/echo-swagger"

	"gorm.io/driver/sqlite"
	"gorm.io/gorm"
)

type Person struct {
	ID        uint   `gorm:"primarykey" json:"id"`
	FirstName string `json:"firstName"`
	LastName  string `json:"lastName"`
	Age       uint   `json:"age"`
}

type API struct {
	DB *gorm.DB
}

// handleIndex godoc
// @Summary  API greeting message
// @Produce  json
// @Success  200
// @Router   / [get]
func (a *API) handleIndex(c echo.Context) error {
	return c.JSON(http.StatusOK, "API v2 Operational")
}

// handlePersonCreate godoc
// @Summary  Create a new Person
// @Accept   json
// @Produce  json
// @Param    firstName  body      string  true  "first name"
// @Param    lastName   body      string  true  "last name"
// @Param    age        body      int     true  "age"
// @Success  202        {object}  Person
// @Router   /person [post]
func (a *API) handlePersonCreate(c echo.Context) error {
	newPerson := new(Person)
	err := json.NewDecoder(c.Request().Body).Decode(&newPerson)
	if err != nil {
		return c.JSON(http.StatusBadRequest, err)
	}

	a.DB.Save(&newPerson)

	return c.JSON(http.StatusCreated, newPerson)
}

// handlePersonGet godoc
// @Summary  Find a registered Person
// @Produce  json
// @Param    firstName  query     string  true  "first name"
// @Success  200        {object}  Person
// @Router   /person/{firstName} [get]
func (a *API) handlePersonGet(c echo.Context) error {
	firstName := c.Param("firstName")

	var person Person
	err := a.DB.Model(&Person{}).
		Where("first_name = ?", firstName).
		Scan(&person).
		Error
	if err != nil {
		return c.JSON(http.StatusNotFound, err)
	}

	return c.JSON(http.StatusCreated, person)
}

// handlePersonUpdate godoc
// @Summary  Update a registered Person
// @Produce  json
// @Param    firstName  query     string  true  "original first name"
// @Param    firstName  body      string  true  "new first name"
// @Param    lastName   body      string  true  "new last name"
// @Param    age        body      int     true  "new age"
// @Success  200        {object}  Person
// @Router   /person/{firstName} [put]
func (a *API) handlePersonUpdate(c echo.Context) error {
	origFirstName := c.Param("firstName")

	updatedPerson := new(Person)
	err := json.NewDecoder(c.Request().Body).Decode(&updatedPerson)
	if err != nil {
		return c.JSON(http.StatusBadRequest, err)
	}

	err = a.DB.Model(&Person{}).
		Where("first_name = ?", origFirstName).
		Updates(*updatedPerson).
		Error
	if err != nil {
		return c.JSON(http.StatusNotFound, err)
	}

	return c.JSON(http.StatusOK, updatedPerson)
}

func createDatabase() (*gorm.DB, error) {
	db, err := gorm.Open(sqlite.Open("people.db"), &gorm.Config{})
	if err != nil {
		return nil, err
	}
	db.AutoMigrate(&Person{})

	return db, nil
}

// @title People API
// @version 2.0

// @BasePath /
// @schemes http
func main() {
	var log = logging.Logger("api")
	var api API
	var err error

	api.DB, err = createDatabase()
	if err != nil {
		log.Fatal("could not create or open database: %s", err)
	}

	e := echo.New()

	e.GET("/", api.handleIndex)
	e.GET("/swagger/*", echoSwagger.WrapHandler)

	e.PUT("/person/:firstName", api.handlePersonUpdate)
	e.POST("/person", api.handlePersonCreate)
	e.GET("/person/:firstName", api.handlePersonGet)

	e.Logger.Fatal(e.Start(":8080"))
}